fix: update schema definition on read
